philadelphia-core: Extract 'FIXConnectionStatusHandler'
In preparation of making it possible for applications to implement
their own administrative message handling, extract the default
administrative message handling out of the 'FIXConnection' class.
